+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+- `genomepy search` now accepts the `--exact` flag
+
 ## [0.15.0] - 2023-02-28
 
 ### Added